pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djoser>=2.3.3",
    "drf-nested-routers>=0.94.2",
    "drf-yasg>=1.21.10",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
